% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdev_function(MDEV).R
\name{MDEV}
\alias{MDEV}
\title{Mean deviation}
\usage{
MDEV(attr)
}
\arguments{
\item{attr}{A vector of numeric values}
}
\value{

}
\description{
Function to calculate mean deviation
Output: A vector of size '1' containing "mean absolute deviation from mean"
}
\examples{
MDEV()
}
