% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdev_distance_function(MDEV_DISTANCE).R
\name{MDEV_DISTANCE}
\alias{MDEV_DISTANCE}
\title{Mean deviation distance}
\usage{
MDEV_DISTANCE(attr)
}
\arguments{
\item{attr}{A vector of numeric values}
}
\value{

}
\description{
Function to calculate mean deviation distances
Output: A vector containing mean deviation distances from each point in given vector
}
\examples{
MDEV_DISTANCE()
}
