% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_outlier_function(RANGE_OUTLIER).R
\name{RANGE_OUTLIER}
\alias{RANGE_OUTLIER}
\title{Tukey range for outlier}
\usage{
RANGE_OUTLIER(attr)
}
\arguments{
\item{attr}{A vector of numeric values}
}
\value{

}
\description{
Function to give outlier ranges based on Tukey's method
Output: A vector of size '2' containing outlier boundaries
}
\examples{
RANGE_OUTLIER()

}
